+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Drop wallet storage format from beta versions
 
 ### Fixed
+- Make device thumbail clickable only in wallet view
 - Improve font weight throughout
 - Improve link appearance in all platforms and change hover cursor
 - Improve 2FA icons and descriptions